requirements.txt

python-dotenv==1.0.0
streamlit==1.26.0
langchain==0.0.318
chromadb==0.4.10
openai==0.27.9
pypdf==3.16.2
tiktoken==0.4.0
